Complete the ProblemCategoryMapper and corresponding unit testing.

+++ web/src/main/java/com/trunkshell/voj/web/mapper/ProblemCategoryMapper.java
@@ -11,35 +11,35 @@ import com.trunkshell.voj.web.model.ProblemCategory;
  */
 @CacheNamespace(implementation = org.mybatis.caches.ehcache.EhcacheCache.class)
 public interface ProblemCategoryMapper {
-	/**
-	 * 通过试题分类的唯一标识符获取试题分类对象.
-	 * @param problemCategoryId - 试题分类的唯一标识符
-	 * @return 预期的试题分类对象或空引用
-	 */
-	public ProblemCategory getProblemCategoryUsingId(long problemCategoryId);
-	
-	/**
-	 * 通过试题分类的唯一英文缩写获取试题分类对象.
-	 * @param problemCategorySlug - 试题分类的唯一英文缩写
-	 * @return 预期的试题分类对象或空引用
-	 */
-	public ProblemCategory getProblemCategoryUsingSlug(String problemCategorySlug);
-	
-	/**
-	 * 创建试题分类对象.
-	 * @param problemCategory - 待创建的试题分类对象
-	 */
-	public void createProblemCategory(ProblemCategory problemCategory);
-	
-	/**
-	 * 更新试题分类对象.
-	 * @param problemCategory - 待更新的试题分类对象
-	 */
-	public void updateProblemCategory(ProblemCategory problemCategory);
-	
-	/**
-	 * 删除试题分类对象.
-	 * @param problemCategoryId - 待删除试题分类对象的唯一标识符
-	 */
-	public void deleteProblemCategory(long problemCategoryId);
+    /**
+     * 通过试题分类的唯一标识符获取试题分类对象.
+     * @param problemCategoryId - 试题分类的唯一标识符
+     * @return 预期的试题分类对象或空引用
+     */
+    public ProblemCategory getProblemCategoryUsingId(int problemCategoryId);
+    
+    /**
+     * 通过试题分类的唯一英文缩写获取试题分类对象.
+     * @param problemCategorySlug - 试题分类的唯一英文缩写
+     * @return 预期的试题分类对象或空引用
+     */
+    public ProblemCategory getProblemCategoryUsingSlug(String problemCategorySlug);
+    
+    /**
+     * 创建试题分类对象.
+     * @param problemCategory - 待创建的试题分类对象
+     */
+    public void createProblemCategory(ProblemCategory problemCategory);
+    
+    /**
+     * 更新试题分类对象.
+     * @param problemCategory - 待更新的试题分类对象
+     */
+    public void updateProblemCategory(ProblemCategory problemCategory);
+    
+    /**
+     * 删除试题分类对象.
+     * @param problemCategoryId - 待删除试题分类对象的唯一标识符
+     */
+    public void deleteProblemCategory(int problemCategoryId);
 }
